+++ builder/build.php
@@ -196,6 +196,9 @@ function build(): bool
         echo "\"$filename\" has been compiled, A total of $progress% has been compiled, ($human_file_index/$total_files).\n";
     }
 
+    sitemap();
+    robots();
+
     logging('built');
 
     return $success;