+++ require.php
@@ -13,6 +13,7 @@ require_once path_join(BUILDER_DIR, 'minimize.php');
 require_once path_join(BUILDER_DIR, 'asset.php');
 require_once path_join(BUILDER_DIR, 'build.php');
 require_once path_join(BUILDER_DIR, 'sitemap.php');
+require_once path_join(BUILDER_DIR, 'robot.php');
 
 // Set global locale
 setlocale(LC_ALL, LOCALE);